+++ fiber/fiber.go
@@ -5,27 +5,47 @@ import (
 
 	"github.com/Becklyn/go-fx-core/env"
 	"github.com/Becklyn/go-fx-core/metrics"
-	"github.com/Becklyn/go-fx-core/middleware"
 	"github.com/gofiber/fiber/v2"
 	fiberlog "github.com/gofiber/fiber/v2/middleware/logger"
 	"github.com/sirupsen/logrus"
 	"go.uber.org/fx"
 )
 
-func newFiber(
-	logger *logrus.Logger,
-	middleware *middleware.MiddlewareRegistry,
-) *fiber.App {
+type MiddlewareHandlerMap map[string][]fiber.Handler
+
+type FiberParams struct {
+	fx.In
+
+	Logger     *logrus.Logger
+	Middleware *MiddlewareHandlerMap `optional:"true"`
+}
+
+func newFiber(params FiberParams) *fiber.App {
 	app := fiber.New()
 
 	app.Use(fiberlog.New(fiberlog.Config{
 		Format: "${latency} - ${status} ${method} ${path}\n",
-		Output: logger.Writer(),
+		Output: params.Logger.Writer(),
 	}))
 
-	app.Use(errorMiddleware(logger))
+	app.Use(errorMiddleware(params.Logger))
 	app.Use(metrics.MetricsMiddleware())
-	app.Use(middleware.Handle)
+
+	if params.Logger != nil {
+		for path, handlers := range *params.Middleware {
+			for _, handler := range handlers {
+				if handler == nil {
+					continue
+				}
+
+				if path == "" {
+					app.Use(handler)
+				} else {
+					app.Use(path, handler)
+				}
+			}
+		}
+	}
 
 	return app
 }