+++ module.go
@@ -6,7 +6,6 @@ import (
 	"github.com/Becklyn/go-fx-core/health"
 	"github.com/Becklyn/go-fx-core/logrus"
 	"github.com/Becklyn/go-fx-core/metrics"
-	"github.com/Becklyn/go-fx-core/middleware"
 	"github.com/Becklyn/go-fx-core/readyness"
 	"go.uber.org/fx"
 )
@@ -17,6 +16,5 @@ var Module = fx.Options(
 	metrics.Module,
 	readyness.Module,
 	health.Module,
-	middleware.Module,
 	fiber.Module,
 )