+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v2.0.0
+
+- (bc) Add `context.Context` as first param in all cqrs handler intarfaces
+- (bc) Replace `MiddlewareRegistry` by `MiddlewareHandlerMap`
+- (bc) Move `cors.Module` to `middleware.NoCors`
+- (internal) Update all dependencies
+
 # v1.1.1
 
 - (bug) Fix module declared as program